+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 3.0.0
+- added localized formatter to format a phone number in an human readable way
 - renamed [dialCode] to [countryCode] as dialCode was semantically incorrect
 
 ## 2.0.1